requirements.txt
torch==2.0.0+cu118
torchvision==0.15.0+cu118
einops
lpips
tqdm
Pillow
numpy<2
setuptools==59.5.0
ninja